\name{getData}

\title{Collect information about the package annotation DB}

\description{
  Return the data stored in AnnotationHub for Myxococcus xanthus DK 1622 in a OrgDb object
}

\usage{
  getData()
}

\arguments{
}

\value{
  \code{getData}: a OrgDb object.
}

\examples{
  getData()
}
\keyword{utilities}
\keyword{datasets}
